nexa/gguf/outetts/requirements.txt
scipy
einops
pyyaml
huggingface_hub
encodec
matplotlib
transformers>=4.46.1
pytorch-lightning
tensorboardX
soundfile
numpy
jsonargparse
torchcrepe
librosa
pesq
inflect
polars
natsort
tqdm
requests
sounddevice
mecab-python3
unidic-lite
uroman